Fix build failure when llvm-core/clang is updated from old version
When we need a new clang version to be built with the same version in
stage1, we need to update seed, but it doesn't work in all profiles,
because llvm-runtimes/clang-runtime for the new version is not updated
when llvm-core/clang compiler is already updated. So clang uses ld by
default, ld wants libgcc, which is missing on musl profiles, and it
fails.  So we use --nodeps and manually manage llvm updates.  It's
better to try to compile new clang with the old one without updating
the seed in stage1, and recompile clang with itself in stage3.

  But today is a full moon, not a day for proper fixes.