Fix llvm-tblgen crashes with OOM on systems without swap
Sometimes 640KB of memory in not enough. But sometimes it's 16GB.
Sometimes there are several different solutions for that and all of
them are sort of... compromise.

It's all about that 721e76e4 (Fix llvm specs, 2025-05-03), all about
that 4b1415ee (Fix llvm and musl specs, 2025-04-11), and that
b8d93ade (Fix build failure when llvm-core/clang is updated from old
version, 2025-05-12).  Shortly AMDGPU backend eats too much memory at
llvm optimizer stage.

As we can't simply enable swap on build systems, we search a
compromise: build with gcc at stage1, disable tmpfs for portage,
disable optimizations. Building with gcc could lead to compiler errors
when llvm libraries not linked properly.  Disabling optimization
dractically slowed clang compiler, and for some profiles (musl) it is
not the solution (not enought memory).  Disabling tmpfs is the
simplest workaround to fix this.  But we also disable optimizations at
stage1 as a workaround when compiler updated from old version to new.

  But today in a new Moon day, not Proper Fixes Day.